add preview generator

+++ Pixels0rtr/MathUtils.swift
@@ -8,13 +8,13 @@
 
 import UIKit
 
-func fMap(value: Float, fromMin: Float, fromMax: Float, toMin: Float, toMax: Float) -> Float {
+func fMap(value: Double, fromMin: Double, fromMax: Double, toMin: Double, toMax: Double) -> Double {
     let m = (toMax - toMin)/(fromMax - fromMin)
     return toMin + m * (value - fromMin)
 }
 
-func fRandom(min: Float, max: Float) -> Float {
+func fRandom(min: Double, max: Double) -> Double {
     let CEIL: UInt32 = 10000
     let r = arc4random_uniform(CEIL)
-    return fMap(value: Float(r), fromMin: 0, fromMax: Float(CEIL), toMin: min, toMax: max)
+    return fMap(value: Double(r), fromMin: 0, fromMax: Double(CEIL), toMin: min, toMax: max)
 }
